Initial set of ports to MinGW64.
Status: Everything in the Makefile up to, but not including, the
definition of EDITOR_OBJS now compiles and runs correctly.

In Makefile, I removed mention of Qt since my goal is to port this to
FLTK.  But I haven't yet touched the source files, so they do not
compile.

I ported the Flex code to flex-2.6.4 on cygwin, more or less.  There is
still some cruft, including the fact that I had to copy FlexLexer.h into
this directory due to it not being otherwise available to the mingw
compiler.

Previously, I had flexlexer.{h,cc} as my own module meant to support
Flex lexers, while FlexLexer.h was something provided by Flex itself.
File names differing only in letter case is questionable in general, and
dangerous on windows, so I renamed mine to bufferlinesource.{h,cc}.  git
is somewhat confused about exactly what happened because I then copied
Flex's FlexLexer.h into this repo.

+++ history.h
@@ -9,6 +9,8 @@
 #include "ogap.h"            // OGapArray
 #include "str.h"             // stringBuilder
 
+#include <stdint.h>          // uintptr_t
+
 
 // fwd in this file
 class HistoryStats;
@@ -155,7 +157,7 @@ class HE_group : public HistoryElt {
 private:     // data
   // single entry in the sequence; see HE_group::encode() in
   // history.cc for encoding
-  typedef unsigned long HistoryEltCode;
+  typedef uintptr_t HistoryEltCode;
 
   // sequence represented as a growable array for efficient
   // storage but also insertion
